requirements.txt
highrise-bot-sdk @ git+https://github.com/itzvini/python-bot-sdk